Top-level in-tree import redirector
For working copy checkout we would like to work on the source
right-away, without putting it under additional wendelin/ (and possibly
/core/) subdirectories.

Add in-tree modules redirector which tweaks plain-python and
pkg_resources-based imports so that e.g.

    import wendelin.bigarray

while in-tree resolves to

    bigarray/__init__.py

In installed state, there will be no such redirector, and code will be
installed with conventional structure.