Hook in list.h from Linux
Unfortunately in CCAN they made separation of list_head into list_head &
list_node (for better type safety I suppose), but that leads e.g. to
inablility to insert element in the middle of the list.

We take the code from util-linux.git (v2.25-165-g9138d6f)[1] which claims
LGPL license for it, with which we are fine.

[1] https://git.kernel.org/cgit/utils/util-linux/util-linux.git/tree/include/list.h?id=v2.25-165-g9138d6f