Fix build with non-std "python"
In 5755a6b3 (Basic setup.py / Makefile to build/install/sdist stuff +
bigfile.so skeleton) we overlooked one thing:

    when tailing from setup.py to make level, we forgot to setup the
    PYTHON env variable properly, and that leads to changing building
    with std python.

Look e.g.

    $ python3 setup.py build_ext -i
    running build_ext
    python setup.py ll_build_ext --inplace
    ^^^

which is obviously not correct.

Fix it.

The tox setup we have setup for testing with multiple interpreters
(7af2b2d7 "Tox setup to test things with py27 & py34 and several
versions of ZODB") did not caught this, because for every python and
environment versions tox initializes separate virtualenv, in which
.../bin/python is a link to chosen python, and this way both python and
my-selected-for-build-python are the same.

Cc: @Tyagov