setup: file_finders entry-point group registration is not a workaround
As discussion in https://bitbucket.org/pypa/setuptools/issue/313
unveiled, setuptools do not have to carry setuptools.file_finders
entrypoint with it in order to support _other_ projects to use
.file_finders entry point.

In our case it means that is is normal that we have to make sure that
the group we are going to register entry-point into, exists.

Remove erroneous comments introduced in 11d130d1 (setup: Ensure
setuptools.file_finders entry-point group is registered).